I2C on SAMW25
Ported the I2C functions from the SAMD21j18 to the SAMW25 board
featuring SAMD21g18. Not much stress. Started some code clean up and
refacturing. Quite sure it can be refactored some more in such a way
that we don't have to bit shift the 16 bits and just pass them to the
I2C and it will split it automaticly. Don't know yet this has to be
tested. The code have not been tested on the board.